requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
mysqlclient==2.2.8
numpy==2.2.6
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
